requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
huggingface_hub
transformers
accelerate
bitsandbytes


-e .